pyproject.toml
[project]
name = "orglog"
version = "0.1.0"
description = "OrgLog logging platform"
authors = [
    {name = "Sujal"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "redis (>=7.2.1,<8.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.28.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "websockets (>=15.0,<16.0)",
    "PyJWT (>=2.8.0,<3.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "python-multipart (>=0.0.9)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = ">=24.1.1"
isort = ">=5.13.2"
pre-commit = ">=3.0.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
